User registration - rename package
