+++ src/main/java/school/faang/BJS243751/Main.java
@@ -1,4 +1,4 @@
-package school.faang.BJS2_43751;
+package school.faang.BJS243751;
 
 public class Main {
     public static void main(String[] args) {
